Update MancalaModel.java I fixed the makeMove method to help with proper errors and ensuring the free turn is working
I add an error message for player already select a pit and making sure the player is suppose to select one pit per turn. I fix some bug in the method. I was add code and fix some if condition for safe running. I was able to add a method from Undomanager class to make sure the player get 3 undo per turn in the makeMove method with clearUndoFlag() and saveUndoState in the right spots.