Update MancalaModel.java I add functions and code for possible actions the player made by misake
I create if condition in the ArrayList<Pit> undo() method by displaying message in case the game is over and the player might press undo which it should show that the game is over and also to make the that the player press the undo button, has not make a move yet. I also fix the free turn in the makemove(int startIdx) method in case the free turn should also have 3 undos. I also make sure if the player and one stone in one of their stone pits and the game is over, it will not display free turn. I added this String.format("Game is still in progress") in getGameResult() method to ensure it follows GUI format